requirements.txt
fastapi
uvicorn
sqlmodel
passlib[bcrypt]
PyJWT
python-dotenv
pytest
httpx